Skeleton code for the subsystems (#13)
* Initial Commit for a drivetrain using VS Code.

* Create Skeleton and add README.md (Closes #2)

Add subsystem classes for Elevator, Intake, and Wrist. Also clean up the
vitruvianlib and add the cheesy spoof factorytalonsrx class to help
configure the talons correctly.

* Cleanup: Switch spaces to tabs to make code style uniform

* Elevator and drive skeleton (#10, #9)

Create the skeleton for the elevator and start working on the skeleton for the
drive. Still need to think through the commands and possibly a full robot state
machine to coordinate between the elevator and wrist

* Start wrist architecture (#11)

Start working on the skeleton for the wrist subsystem, organize the
elevator subsystem a little

* Add Intake skeleton (#12)

Add the roller and clamping functions to the intake